#9 Přidání možnosti odstranit profil, příprava na editaci
