+++ app/src/main/res/menu/menu_profile_detail.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/action_edit"
+        android:title="Upravit"
+        android:icon="@drawable/ic_edit_24"
+        app:showAsAction="ifRoom" />
+    <item
+        android:id="@+id/action_delete"
+        android:title="Smazat profil"
+        android:icon="@drawable/ic_delete_24"
+        app:showAsAction="ifRoom" />
+</menu>
